# development/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Synapse.API | openapi",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/inventory": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "在庫一覧を取得する。itemId・warehouseId でフィルタ可能。\n在庫数量0の行も返す（在庫ゼロの管理目的）。",
        "parameters": [
          {
            "name": "itemId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "warehouseId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/inventory/transactions": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "入出庫履歴を取得する。itemId・warehouseId・日時範囲でフィルタ可能。\n結果は処理日時の降順で返す。",
        "parameters": [
          {
            "name": "itemId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "warehouseId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryTransactionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryTransactionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryTransactionDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/inventory/receive": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "入庫を記録する。\nTransactionType は PurchaseReceipt/ManufacturingReceipt/ReturnReceipt/OtherReceipt のいずれかを指定する。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiveInventoryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiveInventoryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiveInventoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/inventory/issue": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "出庫を記録する。在庫不足の場合は 400 を返す。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueInventoryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueInventoryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IssueInventoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/inventory/adjust": {
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "棚卸調整を記録する。NewQuantity に実地棚卸の実測値を指定する。\n帳簿在庫との差分を自動計算して履歴に記録する。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustInventoryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustInventoryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustInventoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdjustInventoryRequest": {
        "required": [
          "itemId",
          "warehouseId",
          "lotNumber",
          "newQuantity",
          "referenceNumber",
          "note"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "warehouseId": {
            "type": "string",
            "format": "uuid"
          },
          "lotNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "newQuantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "referenceNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "棚卸調整リクエスト。NewQuantity に実地棚卸の実測値を指定する。"
      },
...
      "InventoryTransactionDto": {
        "required": [
          "id",
          "itemId",
          "itemCode",
          "itemName",
          "warehouseId",
          "warehouseCode",
          "warehouseName",
          "lotNumber",
          "transactionType",
          "quantity",
          "referenceNumber",
          "note",
          "transactedAt",
          "createdByUserId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "itemCode": {
            "type": "string"
          },
          "itemName": {
            "type": "string"
          },
          "warehouseId": {
            "type": "string",
            "format": "uuid"
          },
          "warehouseCode": {
            "type": "string"
          },
          "warehouseName": {
            "type": "string"
          },
          "lotNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "transactionType": {
            "$ref": "#/components/schemas/InventoryTransactionType"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "referenceNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          },
          "transactedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdByUserId": {
            "type": "string"
          }
        }
      },
      "InventoryTransactionType": {
        "type": "integer"
      },
      "IssueInventoryRequest": {
        "required": [
          "itemId",
          "warehouseId",
          "lotNumber",
          "quantity",
          "referenceNumber",
          "note"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "warehouseId": {
            "type": "string",
            "format": "uuid"
          },
          "lotNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "referenceNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "出庫リクエスト。"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ReceiveInventoryRequest": {
        "required": [
          "itemId",
          "warehouseId",
          "lotNumber",
          "transactionType",
          "quantity",
          "referenceNumber",
          "note"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "warehouseId": {
            "type": "string",
            "format": "uuid"
          },
          "lotNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "transactionType": {
            "$ref": "#/components/schemas/InventoryTransactionType"
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "referenceNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "入庫リクエスト。"
      },
      "StockDto": {
        "required": [
          "id",
          "itemId",
          "itemCode",
          "itemName",
          "itemUnit",
          "warehouseId",
          "warehouseCode",
          "warehouseName",
          "lotNumber",
          "quantity",
          "lastTransactedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "itemCode": {
            "type": "string"
          },
          "itemName": {
            "type": "string"
          },
          "itemUnit": {
            "type": "string"
          },
          "warehouseId": {
            "type": "string",
            "format": "uuid"
          },
          "warehouseCode": {
            "type": "string"
          },
          "warehouseName": {
            "type": "string"
          },
          "lotNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "quantity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "lastTransactedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}